extracted searchOptions component into its own component file
