+++ blogproject/src/ViewDrafts.js
@@ -4,7 +4,10 @@ import dayjs from 'dayjs'
 import AdvancedFormat from 'dayjs/plugin/advancedFormat'
 // import AdvancedFormat from 'dayjs/plugin/advancedFormat' // ES 2015
 import AlertText from './AlertText'
-import TimePicker from 'react-time-picker'
+import SearchOptions from './SearchOptions'
+
+
+
 import Container from 'react-bootstrap/Container'
 import Table from 'react-bootstrap/Table'
 import Row from 'react-bootstrap/Row'
@@ -423,82 +426,7 @@ const TitleSortOptions = (props)=>{
     )
 }
 
-const SearchOptions = (props)=>{
-const [searchTerm, changeSearchTerm] = useState(undefined)
-const [searchParam, changeSearchParam] = useState(undefined)
 
-const handleSearch =()=>{
-
-const titleSearch = props.array.filter((f)=>{
-    if(f.title.includes(searchTerm)){
-        
-        return f
-    }else{
-        return
-    }
-    
-})
-const articleSearch = props.array.filter((f)=>{
-    if(f.article.includes(searchTerm)){
-        
-        return f
-    }else{
-        return
-    }
-    
-})
-if( !searchTerm ){
-    console.log('search clicked', searchTerm, titleSearch, articleSearch)
-    props.setAlertMessage('No keyword selected')
-    props.setShowAlert(true)
-    return 
-
-}else if(titleSearch.length < 1 || articleSearch.length < 1){
-    props.setAlertMessage('Nothing Found. Try changing the keyword or choosing a different search location from the dropdown')
-    props.setShowAlert(true)
-    return
-}else{
-    console.log('search clicked', searchTerm, titleSearch, articleSearch)
-    props.setShowAlert(false)
-
-   return props.setNewArray(titleSearch)
-
-}
-// takein array and filter by searchTerm
-}
-    return(
-        <Container>
-        <Form.Control type='text' placeholder={`Type keyword(s) here to search VIA ${searchParam}`} onChange={(e)=>{
-            changeSearchTerm(e.target.value)
-        }} />
-        <Button disabled={(searchParam) ? false: true}
-        onClick={()=>{
-            handleSearch()
-        }}
-        >
-        Search
-        </Button>   
-        <Dropdown as={ButtonGroup}>
-  <Button variant="success">{(searchParam )? `Currently Searching via ${searchParam.toUpperCase()} `: `Select search location from dropdown`}</Button>
-
-  <Dropdown.Toggle split variant="success" id="dropdown-split-basic" />
-
-  <Dropdown.Menu>
-    <Dropdown.Item value={'title'} as={Button} onClick={(e)=>{
-        changeSearchParam(e.target.value)
-        console.log('in title button clicked', e.target.value)
-    }}>in title</Dropdown.Item>
-    <Dropdown.Item value={'article'} as={Button} onClick={(e)=>{
-        changeSearchParam(e.target.value)
-        console.log('in article button clicked',e.target.value)
-    }}>in article </Dropdown.Item>
-  </Dropdown.Menu>
-</Dropdown>
-
-
-        </Container>
-    )
-}
